refactor: commands package
