+++ icx-asset/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod list;
+pub mod sync;
