+++ icx-asset/src/support.rs
@@ -1,0 +1,1 @@
+pub type Result<T = ()> = std::result::Result<T, Box<dyn std::error::Error>>;
